+++ src/core/model.jl
@@ -204,10 +204,28 @@ function Base.setproperty!(m::ABM{A, S, F, P}, s::Symbol, x) where {A, S, F, P}
 end
 
 """
-    random_agent(model)
+    random_agent(model) → agent
 Return a random agent from the model.
 """
 random_agent(model) = model[rand(keys(model.agents))]
+
+"""
+    random_agent(model, condition) → agent
+Return a random agent from the model that satisfies `condition(agent) == true`.
+The function generates a random permutation of agent IDs and iterates through them.
+If no agent satisfies the condition, `nothing` is returned instead.
+"""
+function random_agent(model, condition)
+    ids = shuffle!(collect(keys(model.agents)))
+    i, L = 1, length(ids)
+    a = model[ids[1]]
+    while !condition(a)
+        i += 1
+        i > L && return nothing
+        a = model[ids[i]]
+    end
+    return a
+end
 
 """
     nagents(model::ABM)