+++ src/models/predator_prey.jl
@@ -2,19 +2,19 @@ export Sheep, Wolf, Grass
 
 mutable struct Sheep <: AbstractAgent
     id::Int
-    pos::Tuple{Int,Int}
+    pos::Dims{2}
     energy::Float64
 end
 
 mutable struct Wolf <: AbstractAgent
     id::Int
-    pos::Tuple{Int,Int}
+    pos::Dims{2}
     energy::Float64
 end
 
 mutable struct Grass <: AbstractAgent
     id::Int
-    pos::Tuple{Int,Int}
+    pos::Dims{2}
     fully_grown::Bool
     countdown::Int
 end