+++ src/models/sugarscape.jl
@@ -1,6 +1,6 @@
 mutable struct SugarSeeker <: AbstractAgent
     id::Int
-    pos::Tuple{Int,Int}
+    pos::Dims{2}
     vision::Int
     metabolic_rate::Int
     age::Int